OCR based Doc Extractor/Multiple Doc OCR Extractor/requirements.txt
opencv-python
pytesseract
pillow
google-generativeai
streamlit
